# doc/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  BundleManifest:
    properties:
      checksum:
        type: string
        x-go-name: Checksum
      custom_middleware:
        $ref: '#/definitions/MiddlewareSection'
      file_list:
        items:
          type: string
        type: array
        x-go-name: FileList
      signature:
        type: string
        x-go-name: Signature
    type: object
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
...
  File:
    properties:
      Data:
        $ref: '#/definitions/File'
      Header:
        $ref: '#/definitions/FileHeader'
    title: File represents an uploaded file.
    type: object
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/go-openapi/runtime
  FileHeader:
    properties:
      Filename:
        type: string
      Header:
        $ref: '#/definitions/MIMEHeader'
      Size:
        format: int64
        type: integer
    title: A FileHeader describes a file part of a multipart request.
    type: object
    x-go-package: mime/multipart
...
  HookType:
    format: int32
    type: integer
    x-go-package: github.com/TykTechnologies/mserv/coprocess/bindings/go
...
  IdExtractorSource:
    type: string
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
  IdExtractorType:
    type: string
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
...
  MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    description: |-
      A MIMEHeader represents a MIME-style header mapping
      keys to sets of values.
    type: object
    x-go-package: net/textproto
  MW:
    properties:
      APIID:
        type: string
      Active:
        type: boolean
      Added:
        format: date-time
        type: string
      BundleRef:
        type: string
      DownloadOnly:
        type: boolean
      Manifest:
        $ref: '#/definitions/BundleManifest'
      OrgID:
        type: string
      Plugins:
        items:
          $ref: '#/definitions/Plugin'
        type: array
      UID:
        type: string
    type: object
    x-go-package: github.com/TykTechnologies/mserv/storage
...
  MiddlewareDefinition:
    properties:
      name:
        type: string
        x-go-name: Name
      path:
        type: string
        x-go-name: Path
      raw_body_only:
        type: boolean
        x-go-name: RawBodyOnly
      require_session:
        type: boolean
        x-go-name: RequireSession
    type: object
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
  MiddlewareDriver:
    type: string
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
  MiddlewareIdExtractor:
    properties:
      extract_from:
        $ref: '#/definitions/IdExtractorSource'
      extract_with:
        $ref: '#/definitions/IdExtractorType'
      extractor_config:
        additionalProperties:
          type: object
        type: object
        x-go-name: ExtractorConfig
    type: object
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
  MiddlewareSection:
    properties:
      auth_check:
        $ref: '#/definitions/MiddlewareDefinition'
      driver:
        $ref: '#/definitions/MiddlewareDriver'
      id_extractor:
        $ref: '#/definitions/MiddlewareIdExtractor'
      post:
        items:
          $ref: '#/definitions/MiddlewareDefinition'
        type: array
        x-go-name: Post
      post_key_auth:
        items:
          $ref: '#/definitions/MiddlewareDefinition'
        type: array
        x-go-name: PostKeyAuth
      pre:
        items:
          $ref: '#/definitions/MiddlewareDefinition'
        type: array
        x-go-name: Pre
      response:
        items:
          $ref: '#/definitions/MiddlewareDefinition'
        type: array
        x-go-name: Response
    type: object
    x-go-package: github.com/TykTechnologies/mserv/vendor/github.com/TykTechnologies/tyk/apidef
...
  Payload:
    description: Payload is the default response
    properties:
      Error:
        type: string
      Payload:
        type: object
      Status:
        type: string
    type: object
    x-go-package: github.com/TykTechnologies/mserv/models
  Plugin:
    properties:
      FileName:
        type: string
      FileRef:
        type: string
      Name:
        type: string
      Type:
        $ref: '#/definitions/HookType'
      UID:
        type: string
    type: object
    x-go-package: github.com/TykTechnologies/mserv/storage
...
host: localhost:8989
info:
  description: Provides access to operations over an Mserv service.
  title: Mserv API.
  version: 0.0.1
paths:
  /api/mw:
    post:
      description: Expects a file bundle in `uploadfile` form field.
      operationId: mwAdd
      parameters:
      - in: formData
        name: uploadfile
        required: true
        type: file
        x-go-name: UploadFile
      - in: query
        name: store_only
        type: boolean
        x-go-name: StoreOnly
      - in: query
        name: api_id
        type: string
        x-go-name: ApiID
      responses:
        "200":
          $ref: '#/responses/mwIDResponse'
        "500":
          $ref: '#/responses/genericErrorResponse'
      summary: Adds a new middleware. If `store_only` field is true it will only be
        available for download.
      tags:
      - mw
  /api/mw/{id}:
    delete:
      operationId: mwDelete
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/mwIDResponse'
        "500":
          $ref: '#/responses/genericErrorResponse'
      summary: Deletes a middleware specified by {id}.
      tags:
      - mw
    get:
      operationId: mwFetch
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/mwResponse'
        "500":
          $ref: '#/responses/genericErrorResponse'
      summary: Fetches a middleware specified by {id}.
      tags:
      - mw
    put:
      description: Expects a file bundle in `uploadfile` form field.
      operationId: mwUpdate
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: formData
        name: uploadfile
        required: true
        type: file
        x-go-name: UploadFile
      responses:
        "200":
          $ref: '#/responses/mwIDResponse'
        "500":
          $ref: '#/responses/genericErrorResponse'
      summary: Updates a middleware specified by {id}.
      tags:
      - mw
  /api/mw/bundle/{id}:
    get:
      operationId: mwFetchBundle
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      produces:
      - application/octet-stream
      responses:
        "200":
          $ref: '#/responses/mwBundleResponse'
        "500":
          $ref: '#/responses/genericErrorResponse'
      summary: Fetches a middleware bundle file specified by {id}.
      tags:
      - mw
  /api/mw/master/all:
    get:
      operationId: mwListAll
      responses:
        "200":
          $ref: '#/responses/mwListResponse'
        "500":
          $ref: '#/responses/genericErrorResponse'
      summary: Lists all middleware.
      tags:
      - mw
...
produces:
- application/json
responses:
  genericErrorResponse:
    schema:
      $ref: '#/definitions/Payload'
...
  mwBundleResponse:
    schema:
      $ref: '#/definitions/File'
  mwIDResponse:
    schema:
      properties:
        Error:
          type: string
        Payload:
          properties:
            BundleID:
              type: string
          type: object
        Status:
          type: string
      type: object
  mwListResponse:
    schema:
      properties:
        Error:
          type: string
        Payload:
          items:
            $ref: '#/definitions/MW'
          type: array
        Status:
          type: string
      type: object
  mwResponse:
    schema:
      properties:
        Error:
          type: string
        Payload:
          $ref: '#/definitions/MW'
        Status:
          type: string
      type: object
schemes:
- http
- https
securityDefinitions:
  api_key:
    in: header
    name: X-Api-Key
    type: apiKey
swagger: "2.0"